Make code work with FreeRADIUS 3.0.22 and above
This release has gotten an extra argument to the vp_prints_value_json,
to control the translation from integer values to symbolic names. It
would be nice if we could use this and drop our own translation code,
but that would result in a minor behaviour change which we don't want.